mcp-ocr/requirements.txt
tencentcloud-sdk-python
dotenv
fastmcp>=0.2.0
fastapi>=0.115.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
